[MIG] crm_lead_vat: Migration to 12.0
