+++ crm_lead_vat/models/__init__.py
@@ -1,0 +1,1 @@
+from . import crm_lead
