+++ crm_lead_vat/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Rafael Blasco <rafaelbn@antiun.com>
+* Jairo Llopis <yajo.sk8@gmail.com>
