+++ crm_lead_vat/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module was written to extend the functionality of CRM leads to support
+setting the VAT.
