+++ crm_lead_vat/readme/USAGE.rst
@@ -1,0 +1,5 @@
+To use this module, you need to:
+
+* Go to *Sales > Leads*.
+* Open a lead.
+* You will see the new field.
